requirements.txt
streamlit==1.31.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
google-generativeai>=0.7.0
polyline==2.0.0 